# docs/swagger.yaml
basePath: /
definitions:
  dto.CarResponse:
    properties:
      brand:
        example: BMW
        type: string
      color:
        example: Blue
        type: string
      created_at:
        example: "2026-08-06T12:00:00Z"
        type: string
      description:
        type: string
        x-nullable: true
      id:
        example: 23b12e8b-3995-4910-9a7b-55cda03ade9f
        type: string
      image:
        type: string
        x-nullable: true
      model:
        example: X5
        type: string
      model_generation:
        type: string
        x-nullable: true
      price:
        example: 49500
        type: number
      production_year:
        example: 2023
        type: integer
      updated_at:
        example: "2026-08-06T12:00:00Z"
        type: string
    type: object
  dto.CreateCarRequest:
    properties:
      brand:
        example: BMW
        type: string
      color:
        example: Blue
        type: string
      description:
        example: Well maintained. Service records available.
        maxLength: 10000
        type: string
        x-nullable: true
      image:
        example: https://images.example.com/car.jpg
        maxLength: 2048
        type: string
        x-nullable: true
      model:
        example: X5
        type: string
      model_generation:
        example: W124 facelift
        maxLength: 100
        type: string
        x-nullable: true
      price:
        example: 49500
        type: number
      production_year:
        example: 2023
        type: integer
    type: object
  dto.ErrorResponse:
    properties:
      error:
        example: car not found
        type: string
    type: object
  dto.ListCarsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.CarResponse'
        type: array
      limit:
        example: 20
        type: integer
      page:
        example: 1
        type: integer
      total:
        example: 42
        type: integer
    type: object
  generation.Candidate:
    properties:
      body_style:
        type: string
      generation:
        type: string
      production_year_end:
        type: integer
      production_year_start:
        type: integer
      source_url:
        type: string
    type: object
  generation.Result:
    properties:
      candidates:
        items:
          $ref: '#/definitions/generation.Candidate'
        type: array
      notice:
        type: string
    type: object
info:
  contact: {}
  description: API for managing cars.
  title: Car Service API
  version: "1.0"
paths:
  /cars:
    get:
      parameters:
      - description: Page number
        in: query
        minimum: 1
        name: page
        type: integer
      - description: Items per page
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - description: Case-insensitive brand or model search
        in: query
        name: name
        type: string
      - description: Exact production year
        in: query
        maximum: 2100
        minimum: 1886
        name: year
        type: integer
      - description: Alias for year
        in: query
        maximum: 2100
        minimum: 1886
        name: production_year
        type: integer
      - description: Created at or after (RFC3339)
        in: query
        name: created_from
        type: string
      - description: Created at or before (RFC3339)
        in: query
        name: created_to
        type: string
      - description: Exact creation time (RFC3339)
        in: query
        name: created_at
        type: string
      - description: Minimum price
        in: query
        minimum: 0
        name: min_price
        type: number
      - description: Maximum price
        in: query
        minimum: 0
        name: max_price
        type: number
      - description: Exact price
        in: query
        minimum: 0
        name: price
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ListCarsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: List cars
      tags:
      - Cars
    post:
      consumes:
      - application/json
      description: Creates a new car in the database.
      parameters:
      - description: Car data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCarRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CarResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "413":
          description: Request Entity Too Large
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Create car
      tags:
      - Cars
  /cars/{id}:
    delete:
      parameters:
      - description: Car UUID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Delete car
      tags:
      - Cars
    get:
      parameters:
      - description: Car UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CarResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get car
      tags:
      - Cars
    put:
      consumes:
      - application/json
      description: Replaces core fields. Omitted image/model_generation/description
        stay unchanged; null or blank clears them.
      parameters:
      - description: Car UUID
        in: path
        name: id
        required: true
        type: string
      - description: Updated car
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCarRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CarResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "413":
          description: Request Entity Too Large
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Update car
      tags:
      - Cars
  /cars/generations:
    get:
      parameters:
      - description: Manufacturer, as in car CRUD
        in: query
        name: brand
        required: true
        type: string
      - description: Model
        in: query
        name: model
        required: true
        type: string
      - description: Calendar production year
        in: query
        maximum: 2100
        minimum: 1886
        name: year
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/generation.Result'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Suggest model generations (advisory, limited coverage)
      tags:
      - Cars
  /health:
    get:
      description: Reports whether the service is running.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Health check
      tags:
      - System
  /ping:
    get:
      description: Simple ping endpoint
      produces:
      - text/plain
      responses:
        "200":
          description: pong
          schema:
            type: string
      summary: Ping
      tags:
      - System
swagger: "2.0"