# internal/handler/car.go
package handler

import (
	"encoding/json"
	"errors"
	"math"
	"net/http"
	"strconv"
	"strings"
	"time"

	"github.com/InatoInato/car_service.git/internal/db"
	"github.com/InatoInato/car_service.git/internal/handler/dto"
	"github.com/InatoInato/car_service.git/internal/service"
	"github.com/go-chi/chi/v5"
	"github.com/google/uuid"
	"github.com/jackc/pgx/v5"
	"github.com/jackc/pgx/v5/pgtype"
)

type CarHandler struct {
	service *service.CarService
}

func NewCarHandler(service *service.CarService) *CarHandler {
	return &CarHandler{service: service}
}

// Create creates a new car.
//
// @Summary      Create car
// @Description  Creates a new car in the database.
// @Tags         Cars
// @Accept       json
// @Produce      json
// @Param        request body dto.CreateCarRequest true "Car data"
// @Success      201 {object} dto.CarResponse
// @Failure      400 {object} dto.ErrorResponse
// @Failure      500 {object} dto.ErrorResponse
// @Failure      413 {object} dto.ErrorResponse
// @Router       /cars [post]
func (h *CarHandler) Create(w http.ResponseWriter, r *http.Request) {
	req, err := decodeCarRequest(w, r)
	if err != nil {
		h.writeRequestError(w, err)
		return
	}

	params, err := createCarParams(req)
	if err != nil {
		h.writeError(w, http.StatusBadRequest, err.Error())
		return
	}

	car, err := h.service.CreateCar(r.Context(), params)
	if err != nil {
		h.writeError(w, http.StatusInternalServerError, "failed to create car")
		return
	}

	h.writeJSON(w, http.StatusCreated, car)
}

// GetByID returns a car by ID.
//
// @Summary      Get car
// @Tags         Cars
// @Produce      json
// @Param        id path string true "Car UUID"
// @Success      200 {object} dto.CarResponse
// @Failure      400 {object} dto.ErrorResponse
// @Failure      404 {object} dto.ErrorResponse
// @Router       /cars/{id} [get]
func (h *CarHandler) GetByID(w http.ResponseWriter, r *http.Request) {
	idParam := chi.URLParam(r, "id")
	id, err := uuid.Parse(idParam)
	if err != nil {
		h.writeError(w, http.StatusBadRequest, "invalid car id")
		return
	}

	car, err := h.service.GetCarByID(r.Context(), id)
	if err != nil {
		if errors.Is(err, pgx.ErrNoRows) {
			h.writeError(w, http.StatusNotFound, "car not found")
			return
		}
		h.writeError(w, http.StatusInternalServerError, "failed to fetch car")
		return
	}

	h.writeJSON(w, http.StatusOK, car)
}

// List returns all cars.
//
// @Summary      List cars
// @Tags         Cars
// @Produce      json
// @Param        page query int false "Page number" minimum(1)
// @Param        limit query int false "Items per page" minimum(1) maximum(100)
// @Param        name query string false "Case-insensitive brand or model search"
// @Param        year query int false "Exact production year" minimum(1886) maximum(2100)
// @Param        production_year query int false "Alias for year" minimum(1886) maximum(2100)
// @Param        created_from query string false "Created at or after (RFC3339)"
// @Param        created_to query string false "Created at or before (RFC3339)"
// @Param        created_at query string false "Exact creation time (RFC3339)"
// @Param        min_price query number false "Minimum price" minimum(0)
// @Param        max_price query number false "Maximum price" minimum(0)
// @Param        price query number false "Exact price" minimum(0)
// @Success      200 {object} dto.ListCarsResponse
// @Failure      400 {object} dto.ErrorResponse
// @Failure      500 {object} dto.ErrorResponse
// @Router       /cars [get]
func (h *CarHandler) List(w http.ResponseWriter, r *http.Request) {
	params, page, limit, err := listCarsParams(r)
	if err != nil {
		h.writeError(w, http.StatusBadRequest, err.Error())
		return
	}

	cars, total, err := h.service.FilterCars(r.Context(), params)
	if err != nil {
		h.writeError(w, http.StatusInternalServerError, "failed to fetch cars")
		return
	}

	h.writeJSON(w, http.StatusOK, map[string]any{
		"data":  cars,
		"page":  page,
		"limit": limit,
		"total": total,
	})
}

func listCarsParams(r *http.Request) (db.FilterCarsParams, int, int, error) {
	const maxInt32 = int64(1<<31 - 1)

	query := r.URL.Query()
	page := 1
	limit := 20

	if value := query.Get("page"); value != "" {
		parsed, err := strconv.Atoi(value)
		if err != nil || parsed < 1 {
			return db.FilterCarsParams{}, 0, 0, errors.New("invalid page")
		}
		page = parsed
	}

	if value := query.Get("limit"); value != "" {
		parsed, err := strconv.Atoi(value)
		if err != nil || parsed < 1 || parsed > 100 {
			return db.FilterCarsParams{}, 0, 0, errors.New("limit must be between 1 and 100")
		}
		limit = parsed
	}

	if int64(page-1) > maxInt32/int64(limit) {
		return db.FilterCarsParams{}, 0, 0, errors.New("page is too large")
	}

	params := db.FilterCarsParams{
		Name:        strings.TrimSpace(query.Get("name")),
		LimitCount:  int32(limit),
		OffsetCount: int32((page - 1) * limit),
	}

	yearValue := firstQueryValue(r, "year", "production_year")
	if yearValue != "" {
		year, err := strconv.ParseInt(yearValue, 10, 16)
		if err != nil || year < 1886 || year > 2100 {
			return db.FilterCarsParams{}, 0, 0, errors.New("year must be between 1886 and 2100")
		}
		params.Year = pgtype.Int2{Int16: int16(year), Valid: true}
	}

	createdFromValue := query.Get("created_from")
	createdToValue := query.Get("created_to")
	exactCreatedValue := firstQueryValue(r, "created_at", "created_time")
	if exactCreatedValue != "" {
		if createdFromValue != "" || createdToValue != "" {
			return db.FilterCarsParams{}, 0, 0, errors.New("created_at cannot be combined with created_from or created_to")
		}
		createdFromValue = exactCreatedValue
		createdToValue = exactCreatedValue
	}

	createdFrom, err := timestampFilter(createdFromValue)
	if err != nil {
		return db.FilterCarsParams{}, 0, 0, errors.New("created_from must be an RFC3339 timestamp")
	}
	createdTo, err := timestampFilter(createdToValue)
	if err != nil {
		return db.FilterCarsParams{}, 0, 0, errors.New("created_to must be an RFC3339 timestamp")
	}
	if createdFrom.Valid && createdTo.Valid && createdFrom.Time.After(createdTo.Time) {
		return db.FilterCarsParams{}, 0, 0, errors.New("created_from must not be after created_to")
	}
	params.CreatedFrom = createdFrom
	params.CreatedTo = createdTo

	minPriceValue := query.Get("min_price")
	maxPriceValue := query.Get("max_price")
	exactPriceValue := query.Get("price")
	if exactPriceValue != "" {
		if minPriceValue != "" || maxPriceValue != "" {
			return db.FilterCarsParams{}, 0, 0, errors.New("price cannot be combined with min_price or max_price")
		}
		minPriceValue = exactPriceValue
		maxPriceValue = exactPriceValue
	}

	minPrice, minPriceNumber, err := priceFilter(minPriceValue)
	if err != nil {
		return db.FilterCarsParams{}, 0, 0, errors.New("min_price must be a non-negative number")
	}
	maxPrice, maxPriceNumber, err := priceFilter(maxPriceValue)
	if err != nil {
		return db.FilterCarsParams{}, 0, 0, errors.New("max_price must be a non-negative number")
	}
	if minPrice.Valid && maxPrice.Valid && minPriceNumber > maxPriceNumber {
		return db.FilterCarsParams{}, 0, 0, errors.New("min_price must not be greater than max_price")
	}
	params.MinPrice = minPrice
	params.MaxPrice = maxPrice

	return params, page, limit, nil
}

func firstQueryValue(r *http.Request, keys ...string) string {
	for _, key := range keys {
		if value := r.URL.Query().Get(key); value != "" {
			return value
		}
	}
	return ""
}

func timestampFilter(value string) (pgtype.Timestamptz, error) {
	if value == "" {
		return pgtype.Timestamptz{}, nil
	}

	parsed, err := time.Parse(time.RFC3339, value)
	if err != nil {
		return pgtype.Timestamptz{}, err
	}
	return pgtype.Timestamptz{Time: parsed, Valid: true}, nil
}

func priceFilter(value string) (pgtype.Numeric, float64, error) {
	if value == "" {
		return pgtype.Numeric{}, 0, nil
	}

	parsed, err := strconv.ParseFloat(value, 64)
	if err != nil || parsed < 0 || math.IsNaN(parsed) || math.IsInf(parsed, 0) {
		return pgtype.Numeric{}, 0, errors.New("invalid price")
	}

	var numeric pgtype.Numeric
	if err := numeric.Scan(value); err != nil {
		return pgtype.Numeric{}, 0, err
	}
	return numeric, parsed, nil
}

// Update updates a car.
//
// @Summary      Update car
// @Description  Replaces core fields. Omitted image/model_generation/description stay unchanged; null or blank clears them.
// @Tags         Cars
// @Accept       json
// @Produce      json
// @Param        id path string true "Car UUID"
// @Param        request body dto.CreateCarRequest true "Updated car"
// @Success      200 {object} dto.CarResponse
// @Failure      400 {object} dto.ErrorResponse
// @Failure      404 {object} dto.ErrorResponse
// @Failure      413 {object} dto.ErrorResponse
// @Router       /cars/{id} [put]
func (h *CarHandler) Update(w http.ResponseWriter, r *http.Request) {
	idParam := chi.URLParam(r, "id")
	id, err := uuid.Parse(idParam)
	if err != nil {
		h.writeError(w, http.StatusBadRequest, "invalid car id")
		return
	}

	payload, err := decodeCarRequest(w, r)
	if err != nil {
		h.writeRequestError(w, err)
		return
	}

	price, err := numericPrice(payload.Price)
	if err != nil {
		h.writeError(w, http.StatusBadRequest, "invalid price")
		return
	}
	details, err := parseListingDetails(payload)
	if err != nil {
		h.writeError(w, http.StatusBadRequest, err.Error())
		return
	}

	params := db.UpdateCarParams{
		ID:             id,
		Brand:          payload.Brand,
		Model:          payload.Model,
		ProductionYear: payload.ProductionYear,
		Color:          payload.Color,
		Price:          price,
		Image:          details.image, SetImage: payload.Image.Present,
		ModelGeneration: details.generation, SetModelGeneration: payload.ModelGeneration.Present,
		Description: details.description, SetDescription: payload.Description.Present,
	}

	car, err := h.service.UpdateCar(r.Context(), params)
	if err != nil {
		if errors.Is(err, pgx.ErrNoRows) {
			h.writeError(w, http.StatusNotFound, "car not found")
			return
		}
		h.writeError(w, http.StatusInternalServerError, "failed to update car")
		return
	}

	h.writeJSON(w, http.StatusOK, car)
}

func createCarParams(req dto.CreateCarRequest) (db.CreateCarParams, error) {
	price, err := numericPrice(req.Price)
	if err != nil {
		return db.CreateCarParams{}, errors.New("invalid price")
	}
	details, err := parseListingDetails(req)
	if err != nil {
		return db.CreateCarParams{}, err
	}

	now := time.Now().UTC()
	return db.CreateCarParams{
		ID:              uuid.New(),
		Brand:           req.Brand,
		Model:           req.Model,
		ProductionYear:  req.ProductionYear,
		Color:           req.Color,
		Price:           price,
		Image:           details.image,
		ModelGeneration: details.generation,
		Description:     details.description,
		CreatedAt:       pgtype.Timestamptz{Time: now, Valid: true},
		UpdatedAt:       pgtype.Timestamptz{Time: now, Valid: true},
	}, nil
}

func numericPrice(price float64) (pgtype.Numeric, error) {
	var numeric pgtype.Numeric
	if err := numeric.Scan(strconv.FormatFloat(price, 'f', -1, 64)); err != nil {
		return pgtype.Numeric{}, err
	}
	return numeric, nil
}

// Delete removes a car.
//
// @Summary      Delete car
// @Tags         Cars
// @Param        id path string true "Car UUID"
// @Success      204
// @Failure      400 {object} dto.ErrorResponse
// @Failure      500 {object} dto.ErrorResponse
// @Router       /cars/{id} [delete]
func (h *CarHandler) Delete(w http.ResponseWriter, r *http.Request) {
	idParam := chi.URLParam(r, "id")
	id, err := uuid.Parse(idParam)
	if err != nil {
		h.writeError(w, http.StatusBadRequest, "invalid car id")
		return
	}

	err = h.service.DeleteCar(r.Context(), id)
	if err != nil {
		h.writeError(w, http.StatusInternalServerError, "failed to delete car")
		return
	}

	w.WriteHeader(http.StatusNoContent)
}

func (h *CarHandler) writeJSON(w http.ResponseWriter, status int, data any) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(status)
	_ = json.NewEncoder(w).Encode(data)
}

func (h *CarHandler) writeError(w http.ResponseWriter, status int, msg string) {
	h.writeJSON(w, status, map[string]string{"error": msg})
}